requirements.txt
fastapi==0.110.2
uvicorn==0.30.1
pydantic==1.10.15
sentence-transformers==2.6.1
transformers==4.41.2
tokenizers==0.19.1
torch==2.2.2
pinecone-client==5.0.1
requests==2.32.2
python-dotenv==1.0.1
pyyaml==6.0.1
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.12.0
